+++ ios/sdk/component/image/UIImageView+Hippy.h
