+++ ios/sdk/component/modal/HippyModalHostView.m
@@ -27,7 +27,7 @@
 #import "HippyModalHostViewController.h"
 #import "HippyTouchHandler.h"
 #import "HippyUIManager.h"
-#import "UIView+React.h"
+#import "UIView+Hippy.h"
 
 #import <UIKit/UIKit.h>
