+++ ios/sdk/component/textinput/HippyTextView.mm
@@ -27,7 +27,7 @@
 #import "HippyText.h"
 #import "HippyUtils.h"
 #import "HippyTextSelection.h"
-#import "UIView+React.h"
+#import "UIView+Hippy.h"
 
 @implementation HippyUITextView
 {
