+++ ios/sdk/component/view/HippyViewManager.mm
@@ -31,7 +31,7 @@
 #import "HippyUIManager.h"
 #import "HippyUtils.h"
 #import "HippyView.h"
-#import "UIView+React.h"
+#import "UIView+Hippy.h"
 #import "HippyVirtualNode.h"
 #import "HippyConvert+Transform.h"
